homework4/requirements.txt
torch
pillow
transformers
requests
python-dotenv
boto3